+++ frontend/src/components/Navigation/index.jsx
@@ -5,8 +5,8 @@ import logo from "../../static/images/logo/logo.svg";
 class Navigation extends Component {
     render() {
         return (
-            <nav className="navbar navbar-expand-md navbar-light fixed-top bg-jungle">
-                <a className="navbar-brand" href="/"><img src={logo} className="logo"/></a>
+            <nav className="navbar navbar-expand-sm navbar-light fixed-top bg-jungle">
+                <a className="navbar-brand" href="/"><img src={logo} alt="logo" className="logo"/></a>
                 <button className="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
                     <span className="navbar-toggler-icon"></span>
                 </button>