+++ frontend/src/features/HomePage/paths.js
@@ -1,5 +1,0 @@
-export const HOME = `/`;
-export const WELCOME = `/welcome`;
-export const LOGIN = `${WELCOME}/login`;
-
-