+++ frontend/src/features/entrance/paths.js
@@ -1,0 +1,5 @@
+export const FEED = `/`;
+export const WELCOME = `/welcome`;
+export const SANDBOX = `/sandbox`;
+export const PROFILE = `/profile`;
+