+++ frontend/src/helpers/index.js
@@ -1,4 +1,4 @@
 export * from './fake_backend';
 export * from './history';
-export * from './store';
+export * from '../store/store';
 export * from './auth-header';