+++ frontend/src/index.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
+import { Router } from 'react-router-dom';
 
-import { store } from './helpers/store';
+import { store } from './store/store';
 
 
 import 'bootstrap-css-only/css/bootstrap.min.css';
@@ -10,7 +11,7 @@ import 'bootstrap/dist/js/bootstrap.js';
 // import $ from 'jquery';
 // import Popper from 'popper.js';
 
-import './index.scss'
+import './index.scss';
 
 import {App} from './App';
 import * as serviceWorker from './serviceWorker';
@@ -19,7 +20,6 @@ import * as serviceWorker from './serviceWorker';
 // setup fake backend
 import { configureFakeBackend } from './helpers/fake_backend';
 configureFakeBackend();
-
 
 ReactDOM.render(
     <Provider store={store}>