+++ frontend/src/utils/date.js
@@ -1,0 +1,13 @@
+const days = ["Sunday","Monday","Tuesday","Wednesday","Thursday","Friday","Saturday"];
+const months = ["January","February","March","April","May","June","July","August","September","October","November","December"];
+
+export const getDay = (date) => {
+    const d = new Date(date);
+    let day = days[d.getDay()];
+    let month = months[d.getMonth()];
+
+    return(`${day}, ${month} ${d.getDate()}`);
+};
+
+
+