+++ frontend/CHANGELOG.md
@@ -6,5 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
--  [FE-00](https://github.com/lysnikolaou/eatandmeet/issues/10) React-Framework via Create-React-App
--  [FE-01](https://github.com/lysnikolaou/eatandmeet/issues/5) Landing - Login - Registration Page
+-   [FE-00](https://github.com/lysnikolaou/eatandmeet/issues/10) React-Framework via Create-React-App
+-   [FE-01](https://github.com/lysnikolaou/eatandmeet/issues/5) Landing - Login - Registration Page
+-   [FE-04](https://github.com/lysnikolaou/eatandmeet/issues/8) User Profile Page
+-   [FE-11](https://github.com/lysnikolaou/eatandmeet/issues/17) Events Feed